feat(js/api): rename useDatasetDependencies as useAssetDependencies
