+++ airflow/www/static/js/datasets/SearchBar.tsx
@@ -20,7 +20,7 @@
 import React from "react";
 import { Select, SingleValue, useChakraSelectProps } from "chakra-react-select";
 
-import type { DatasetDependencies } from "src/api/useDatasetDependencies";
+import type { DatasetDependencies } from "src/api/useAssetDependencies";
 import type { OnSelectProps } from "./types";
 
 interface Props {
